feat: ✨ rebranding and break layout and index files into components
