+++ web-app/src/components/Footer.tsx
@@ -1,0 +1,5 @@
+import { Box } from '@chakra-ui/react'
+
+export const Footer: React.FC = () => {
+  return <Box></Box>
+}
