+++ web-app/src/styles/theme.ts
@@ -13,8 +13,8 @@ const theme = extendTheme({
       900: '#020024',
       800: '#1a0a3f',
       700: '#3a1765',
-      600: '#4a1e78',
-      500: '#612893',
+      600: '#1D2C57',
+      500: '#5870B3',
       400: '#7746a2',
       300: '#8a5faf',
       200: '#a584c2',
